requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-filter==24.3
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2
